Refactor internal recorder into dedicated recording classes
