+++ Application/Recording/HardwareAccelerationApi.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+namespace ToNRoundCounter.Application.Recording
+{
+    internal enum HardwareAccelerationApi
+    {
+        Auto,
+        Direct3D11,
+        Software,
+    }
+}
